rename and refactor cache functions
