--- util/activity-report/index.js
@@ -1,5 +1,0 @@
-const { queryDeveloperActivity } = require('./developer-activity.js');
-
-module.exports = {
-  queryDeveloperActivity,
-};
