--- util/api-config.js
@@ -1,5 +1,0 @@
-const avgTxCost = require('./activity-report/avg-tx-cost.js');
-const allActivity = require('./activity-report/all-activity.js');
-const devActivity = require('./activity-report/developer-activity.js');
-
-module.exports = [avgTxCost, allActivity, devActivity];
